+++ examples/Example250_NSELidDrivenCavity.jl
@@ -21,25 +21,62 @@ The computed solution for the default parameters looks like this:
 module Example250_NSELidDrivenCavity
 
 using ExtendableFEM
-using GridVisualize
+using Triangulate
 using ExtendableGrids
+using SimplexGridFactory
 using LinearAlgebra
+using GridVisualize
+using LinearSolve
+using IncompleteLU
 using Test #hide
 
-function kernel_nonlinear!(result, u_ops, qpinfo)
-    u, ∇u, p = view(u_ops, 1:2), view(u_ops, 3:6), view(u_ops, 7)
+
+function create_cone(h)
+    builder = SimplexGridBuilder(; Generator = Triangulate)
+
+    ## points
+    p1 = point!(builder, -1, 0)
+    p2 = point!(builder, 1, 0)
+    p3 = point!(builder, 0, -2)
+
+    ## top face
+    facetregion!(builder, 1)
+    facet!(builder, p1, p2)
+
+    ## other faces
+    facetregion!(builder, 2)
+    facet!(builder, p2, p3)
+    facet!(builder, p3, p1)
+
+    cellregion!(builder, 1)
+    maxvolume!(builder, h)
+    regionpoint!(builder, 0, -0.5)
+
+    return simplexgrid(builder)
+end
+
+const 𝕀 = [1 0; 0 1]
+
+function NSE_kernel!(result, u_ops, qpinfo)
+
+    u = tensor_view(u_ops, 1, TDVector(2))
+    v = tensor_view(result, 1, TDVector(2))
+    ∇u = tensor_view(u_ops, 3, TDMatrix(2))
+    ∇v = tensor_view(result, 3, TDMatrix(2))
+    p = tensor_view(u_ops, 7, TDScalar())
+    q = tensor_view(result, 7, TDScalar())
     μ = qpinfo.params[1]
-    result[1] = dot(u, view(∇u, 1:2))
-    result[2] = dot(u, view(∇u, 3:4))
-    result[3] = μ * ∇u[1] - p[1]
-    result[4] = μ * ∇u[2]
-    result[5] = μ * ∇u[3]
-    result[6] = μ * ∇u[4] - p[1]
-    result[7] = -(∇u[1] + ∇u[4])
+
+    tmul!(v, ∇u, u)
+    ∇v .= μ .* ∇u .- p[1] .* 𝕀
+    q[1] = -dot(∇u, 𝕀)
+
     return nothing
 end
 
-function boundarydata!(result, qpinfo)
+
+## boundary function
+function u_boundary!(result, qpinfo)
     result[1] = 1
     result[2] = 0
     return nothing
@@ -58,7 +95,16 @@ function initialgrid_cone()
     return xgrid
 end
 
-function main(; μ_final = 0.0005, order = 2, nrefs = 5, Plotter = nothing, kwargs...)
+function main(;
+        μ_final = 0.0005,        # flow parameter
+        order = 2,               # FE order of the flow field (pressure order is order-1)
+        h = 1.0e-3,              # grid cell volume
+        nrefs = 1,               # additional grid refinements
+        method_linear = nothing, # linear solver ("nothing" invokes the default solver)
+        precon_linear = nothing, # preconditioner
+        Plotter = nothing,
+        kwargs...
+    )
 
     ## prepare parameter field
     extra_params = Array{Float64, 1}([max(μ_final, 0.05)])
@@ -70,30 +116,52 @@ function main(; μ_final = 0.0005, order = 2, nrefs = 5, Plotter = nothing, kwar
 
     assign_unknown!(PD, u)
     assign_unknown!(PD, p)
-    assign_operator!(PD, NonlinearOperator(kernel_nonlinear!, [id(u), grad(u), id(p)]; params = extra_params, kwargs...))
-    assign_operator!(PD, InterpolateBoundaryData(u, boundarydata!; regions = 3))
-    assign_operator!(PD, HomogeneousBoundaryData(u; regions = [1, 2]))
-    assign_operator!(PD, FixDofs(p; dofs = [1]))
+    assign_operator!(PD, NonlinearOperator(NSE_kernel!, [id(u), grad(u), id(p)]; params = extra_params, kwargs...))
 
+    ## boundary data
+    assign_operator!(PD, InterpolateBoundaryData(u, u_boundary!; regions = 1))
+    assign_operator!(PD, HomogeneousBoundaryData(u; regions = [2]))
 
     ## grid
-    xgrid = uniform_refine(initialgrid_cone(), nrefs)
+    xgrid = uniform_refine(create_cone(h), nrefs)
 
     ## prepare FESpace
-    FES = [FESpace{H1Pk{2, 2, order}}(xgrid), FESpace{H1Pk{1, 2, order - 1}}(xgrid)]
+    FES = [
+        FESpace{H1Pk{2, 2, order}}(xgrid),
+        FESpace{H1Pk{1, 2, order - 1}}(xgrid),
+    ]
 
     ## prepare plots
     plt = GridVisualizer(; Plotter = Plotter, layout = (1, 2), clear = true, size = (1600, 800))
 
+
+    ## prepare an initial solution matching the boundary data
+    sol = FEVector(FES; tags = PD.unknowns)
+    interpolate!(sol[u], ON_BFACES, u_boundary!; regions = [1])
+
+
     ## solve by μ embedding
     step = 0
-    sol = nothing
     SC = nothing
     PE = PointEvaluator([id(1)])
     while (true)
         step += 1
         @info "Step $step : solving for μ=$(extra_params[1])"
-        sol, SC = ExtendableFEM.solve(PD, FES, SC; return_config = true, target_residual = 1.0e-10, maxiterations = 20, kwargs...)
+        sol, SC = ExtendableFEM.solve(
+            PD,
+            FES,
+            SC;
+            init = sol,
+            method_linear,
+            # use new preconditioner API: https://docs.sciml.ai/LinearSolve/stable/basics/Preconditioners/#Specifying-Preconditioners
+            # this does currently now work with ILU, see https://github.com/WIAS-PDELib/ExtendableFEM.jl/pull/47#issuecomment-2796849931
+            precon_linear,
+            return_config = true,
+            target_residual = 1.0e-10,
+            maxiterations = 20,
+            kwargs...
+        )
+
         if step == 1
             initialize!(PE, sol)
         end
@@ -116,9 +184,18 @@ function main(; μ_final = 0.0005, order = 2, nrefs = 5, Plotter = nothing, kwar
 end
 
 generateplots = ExtendableFEM.default_generateplots(Example250_NSELidDrivenCavity, "example250.png") #hide
-function runtests() #hide
-    sol, plt = main(; nrefs = 3, μ_final = 0.005) #hide
-    @test sum(view(sol[1])) ≈ 9.501630403050289 #hide
-    return nothing #hide
-end #hide
+function runtests()                                                                                  #hide
+    sol, plt = main(; μ_final = 0.005)                                                               #hide
+    sum(view(sol[1])) ≈ 237.24628017878518                                                           #hide
+
+    method_linear = KrylovJL_GMRES()                                                                 #hide
+    precon_linear = IncompleteLU.ilu                                                                 #hide
+
+    sol, plt = main(; μ_final = 0.005, method_linear, precon_linear)                                 #hide
+    @test sum(view(sol[1])) ≈ 237.24628017878518                                                     #hide
+
+    return nothing                                                                                   #hide
+end                                                                                                  #hide
+
+
 end # module